python-backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
Pillow>=9.0.0
pytesseract>=0.3.10
PyMuPDF>=1.23.0
pdfplumber>=0.9.0
reportlab>=4.0.0
matplotlib>=3.6.0
opencv-python-headless>=4.7.0
requests>=2.28.0
python-dotenv>=1.0.0
openai>=1.0.0
aiofiles>=23.0.0
numpy>=1.20.0
